+++ src/components/cards/FinanceInfo/LargeCard/index.tsx
