+++ src/components/cards/FinanceInfo/SmallCard/index.tsx
